requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
twilio==8.11.0
APScheduler==3.10.4
streamlit==1.29.0
pandas==2.1.4
plotly==5.18.0
requests==2.31.0
pytest==7.4.3
pytest-flask==1.3.0
